More point quality of life improvements.
1. Add a point key that can be used in Maps. Add a test.
2. Move SGF logic to a separate file
3. Add an equals method for points.

Fixes: #196
